+++ .env.example
@@ -1,0 +1,10 @@
+JAWSDB_MARIA_URL="mysql://xxxxxxxxxxxxxxxxxx:xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx@exxxxxxxxxxxxxxxxxxxx.compute-1.amazonaws.com:xxxx/xxxxxxxxxxxxxxxxxxxxxxxx"
+DB_DBNAME_LOCAL=<your database>
+DB_USER_LOCAL=<your database user>
+DB_PASSWORD_LOCAL=<your database passowrd>
+DB_HOST_LOCAL=<your database host>
+DB_PORT_LOCAL=<your database port>
+EMAIL_BOT="xxxxxxxxxxxxxxxxxxxxxxxxxx@gmail.com"
+EMAIL_BOT_CLIENT_ID='xxxxxxxxxxxxxxxxxxxxxxxx-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx.apps.googleusercontent.com'
+EMAIL_BOT_CLIENT_SECRET='xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
+EMAIL_BOT_REFRESH_TOKEN='1//xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
