--- backend/.env.example
@@ -1,6 +1,0 @@
-JAWSDB_MARIA_URL="mysql://xxxxxxxxxxxxxxxxxx:xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx@exxxxxxxxxxxxxxxxxxxx.compute-1.amazonaws.com:xxxx/xxxxxxxxxxxxxxxxxxxxxxxx"
-DB_DBNAME_LOCAL=<your database>
-DB_USER_LOCAL=<your database user>
-DB_PASSWORD_LOCAL=<your database passowrd>
-DB_HOST_LOCAL=<your database host>
-DB_PORT_LOCAL=<your database port>
