--- backend/gdrive/credentials.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "dagama",
-  "private_key_id": "5b40aa7890594adc142f28ddd76fee69ae595088",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQCu2XW9CYM9YjE4\nKwJCxC2waxFP3sBUbYzMdTvAfqnutQqPTvI+atVPh7+xE0n3H/u44Y4L4xQ799Ax\nV/mHtb6yl1pvvJCRsYT+RbK0WqlwG9E6/y0bMykiyBMrkXDnqRNB3JUjv8Dpuwwp\n/sux/shNGzR6R3fbOvFzqqzK5teju+p7kMzeYvVOHRf1ntLSGA71loRx91gbOQGw\nn52QynIVF8j1tcFOMiHTd4QZH620PsdVFOo5zp/qYu3vzDN6pLF3GgJubTjTiIvl\nEaySXaYlN5xMxovsGdGWSwIzHI6r+A4pTjbUeFVHxhfV4wV8qvlEhftvgLRTiFy5\niOb+Q7rjAgMBAAECggEABq8+npMCBvM9kXmndPj+DCwTtdt8u92ST/C9b48IHR2j\nwTXR8gvwisFUVUc5AS8iVv+ciBzO1boUJgbIfeU3XKmHphTt6/v+6x+sBNQgfS47\nEgfWaB9NF8zRWs2NJPYkS9TMiNjYXbcucF6S2AO3EaeUGoan8OQwfe0M/RbVMKe6\nwbPVJ1GHgIdDQV/zeXmUhvpzaAbH5+8/mmzdwcmYb47PwFVvVtP4Ft6tut1scjfh\nHD/Yvdcx1Bq9Xa/Ug2NqruGyLUJ1UgsbOY1tmhHtXpqtTRp5ID2vcIrsyKhUx8pN\noaOzkm42jjFwAuX+poYz4Oz6ZfNdIQzSyJwLl0XiAQKBgQDnbVkBzsEmvCn0uvjj\nss4jpJGDNxFobQg+4pu/Qst578BGwZ/8bdnaTjSTBdH015+wul0qCET85mh18oIV\ndHLZ6xMree5FXzOoua456Cbe0pzxKt9Q5gipYvAcJftF0erAwgXviFlqAKWLFzhu\nH+DNtlmBRD5zUgMaYU8DoT82AQKBgQDBaje0q1QR5c79W++rO/6QlDO17ePZOyAa\nP70/JqxwpVeURgyBKI0cxtSiI/5Z7TEQJ3fxeVJ+990Yqt18Dd1wHXxIu1CoTevS\n6Et42MB3Q+LdbNIbok/s3VZenA+3fr/2trCG1juHkH1a5mlzFpxAlyDzDK6AG6GU\nJIr+qabY4wKBgGoUHTBO7QosRa7/5QW59/Mug+R5YWlUUOtCQ9GvdEU4OBI2vnae\ng7UlE8QwbsSj8Xa4devGTpj9o/QsbKQY9Gwe7HKJOUZjFPLJ6apxVx/bukRxOluC\nlYXO6/B7GGowhkJErSPWg+Wq1jajNiKESi8A3pX6CgYiGXoGEInBZ5ABAoGAKHGV\nhGnvAm8TgVcgHqZE7OaQiZJc+qeqZtK9N99WX5ezYIXlR3y7RTEudVkhbpxyjfYy\n1oOQBIy/5tCTz8T2cOM5gdvpWcgxb8CzR2nLlSTU44hFxxGBtP3zeYUYpD4U1foN\nn6RcflRwYfmcQzyPH9J6SVh1hMRCXT/m8Kl5AhECgYANjq87K+nZSyjKxmmsD5x1\nBulU2aOqVr/XTTlVeadin3tCYYyRLMQVs+s68t27E2boyCkouAIgqPjPPpUuuoZO\nVDPA0I/PVTzynPDrUebu2JOot0KSsKSAXxWI+IITrc95OW52mTonTvcijC1J15gH\nDJ7u2enR+KhtxW0DLyzV0w==\n-----END PRIVATE KEY-----\n",
-  "client_email": "dagama-gdrive@dagama.iam.gserviceaccount.com",
-  "client_id": "111199218700182368201",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/dagama-gdrive%40dagama.iam.gserviceaccount.com"
-}
