+++ composer.json
@@ -5,7 +5,8 @@
         "phpmailer/phpmailer": "^6.5",
         "league/oauth2-google": "^4.0",
         "google/apiclient": "^2.0",
-        "vlucas/phpdotenv": "^5.4"
+        "vlucas/phpdotenv": "^5.4",
+        "pecee/simple-router": "4.3.7.2"
     },
     "license": "MIT",
     "autoload": {
