+++ config/gdrive/driver.php
@@ -1,10 +1,5 @@
 <?php
-  if(preg_match("/localhost/","$_SERVER[HTTP_HOST]")){
-    require_once '../vendor/autoload.php';
-  }
-  else{
-    require_once '/app/vendor/autoload.php';
-  }
+
   function getClient(){
       $client = new Google\Client();
       $GOOGLE_OAUTH_CREDENTIALS_PATH=NULL;
@@ -40,7 +35,7 @@
                 printf("%s (%s)\n<br>", $file->getName(), $file->getId());
             }
         }
-  }; 
+  }
   function insertFile($mimeType, $filename,$folderID,$newfilename) {
       $client=getClient();
       $service=new Google\Service\Drive($client);      
@@ -79,37 +74,7 @@
     catch(Exception $e){
       return $e->getMessage();
     }
-  };
-  /* 
-  SEM PERMISSÂO;
-    function getFile($file_id){
-      $client=getClient();
-      $service=new \Google\Service\Drive($client);
-      try{
-        $result=$service->files->get($file_id);
-        echo "<br>";
-        echo '<pre>' , var_dump($result) , '</pre>';
-        return $result;
-      }
-      catch(Exception $e){
-        return false;
-      }
-    };
-  */
-  //getAllfiles();
-
-  /*
-  //testando
-  if(isset($_FILES["foto"])){
-    $file=$_FILES["foto"];
-    $server_path=$file["tmp_name"];
-    $filename=$file["name"];
-    $filetype=$file["type"];
-    echo "PATH:$server_path<br>";
-    echo "FILE:$filename<br>";
-    echo "TYPE:$filetype<br>";
-    insertFile("$filetype","$server_path",$FOLDERS['avatares'],"$filename");
   }
-  */
+
 ?>
