+++ database/services.php
@@ -1,0 +1,12 @@
+<?php
+    require '../config/gdrive/driver.php';
+    require '../config/mailer/mailer.php';
+    require '../database/database.php';
+    require '../controllers/AssuntoController.php';
+    require '../controllers/FriendController.php';
+    require '../controllers/LocalController.php';
+    require '../controllers/Login_RegisterController.php';
+    require '../controllers/PortoController.php';
+    require '../controllers/PostController.php';
+    require '../controllers/UserController.php';
+?>
