--- index.php
@@ -1,4 +1,0 @@
-<?php
-    header("Location: public/");
-    die();
-?>
