+++ public/view/amigos.php
